+++ es/index.js
@@ -1,5 +1,6 @@
 import startup from './core/startup';
 import register from './core/register';
+import r from './core/r';
 import registerToDefault from './core/register-to-default';
 import registerSingleClassToDefault from './core/register-single-class-to-default';
 import configure from './core/configure';
@@ -7,10 +8,17 @@ import invoke from './core/invoke';
 import invokeSingle from './core/invoke-single';
 import setGlobalState from './core/set-global-state';
 import setState from './core/set-state';
+import emit from './core/emit';
+import emitWith from './core/emit-with';
+import off from './core/off';
 import ccContext from './cc-context';
 var defaultExport = {
+  emit: emit,
+  emitWith: emitWith,
+  off: off,
   startup: startup,
   register: register,
+  r: r,
   registerToDefault: registerToDefault,
   registerSingleClassToDefault: registerSingleClassToDefault,
   configure: configure,